node-data: move faker methods to relevant modules
